refactored into OOP Command pattern
