+++ Explorer/Assets/DCL/Chat/ChatCommands/IChatCommand.cs
@@ -1,0 +1,9 @@
+﻿using Cysharp.Threading.Tasks;
+
+namespace DCL.Chat.ChatCommands
+{
+    internal interface IChatCommand
+    {
+        UniTask<string> ExecuteAsync();
+    }
+}
