+++ Explorer/Assets/DCL/Chat/ChatCommands/IChatCommand.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 68ce779b5a074a26a7640ba14f38ba2b
+timeCreated: 1712245170
