+++ Explorer/Assets/DCL/Chat/ChatCommands/TeleportToChatCommand.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 3b9cc816ec6e4659961ec42134e20571
+timeCreated: 1712245241
